# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "PhotoSet 摄影套图浏览平台后端 API 文档\n提供套图管理、社区互动、用户认证、支付等功能",
        "title": "PhotoSet API",
        "termsOfService": "https://photoset.example.com/terms",
        "contact": {
            "name": "API Support",
            "email": "support@photoset.example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/api/photosets/{id}/reviews": {
            "get": {
                "description": "分页获取套图评价列表，支持排序",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Review"
                ],
                "summary": "获取评价列表",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "套图ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "每页数量",
                        "name": "page_size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "default": "newest",
                        "description": "排序方式(newest/highest/lowest)",
                        "name": "sort_by",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "评价列表",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "对套图进行评分（1-5星）和评价",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Review"
                ],
                "summary": "创建评价",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "套图ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "评价请求 {rating, tags, content, is_anonymous}",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "评价结果",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/api/photosets/{id}/reviews/mine": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "获取当前用户对指定套图的评价",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Review"
                ],
                "summary": "获取我的评价",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "套图ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "我的评价",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "401": {
                        "description": "未登录",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/api/photosets/{id}/reviews/summary": {
            "get": {
                "description": "获取套图的评价统计汇总（评分分布、标签统计等）",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Review"
                ],
                "summary": "获取评价汇总",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "套图ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "评价汇总",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/api/photosets/{id}/reviews/{reviewId}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "更新自己对套图的评价",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Review"
                ],
                "summary": "更新评价",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "套图ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "评价ID",
                        "name": "reviewId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "更新请求 {rating, tags, content, is_anonymous}",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "更新结果",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "删除自己对套图的评价",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Review"
                ],
                "summary": "删除评价",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "套图ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "评价ID",
                        "name": "reviewId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "删除成功",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
...
    },
...
    "securityDefinitions": {
        "BearerAuth": {
            "description": "在请求头中添加 Bearer {token} 进行认证",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}